refactor: group table page components
