+++ apps/desktop/src/entities/connection/components/table/table-cell-table.tsx
@@ -10,11 +10,11 @@ import { useMemo } from 'react'
 import { Table, TableBody, TableHeader, TableProvider } from '~/components/table'
 import { DEFAULT_COLUMN_WIDTH, DEFAULT_ROW_HEIGHT, getDisplayValue } from '~/entities/connection/components/table/utils'
 import { connectionRowsQuery } from '~/entities/connection/queries'
-import { TableError } from '~/routes/_protected/database/$id/table/-components/table'
-import { TableEmpty } from '~/routes/_protected/database/$id/table/-components/table-empty'
-import { TableHeaderCell } from '~/routes/_protected/database/$id/table/-components/table-header-cell'
-import { TableInfiniteLoader } from '~/routes/_protected/database/$id/table/-components/table-infinite-loader'
-import { TableBodySkeleton } from '~/routes/_protected/database/$id/table/-components/table-skeleton'
+import { TableError } from '~/routes/_protected/database/$id/table/-components/table/table'
+import { TableEmpty } from '~/routes/_protected/database/$id/table/-components/table/table-empty'
+import { TableHeaderCell } from '~/routes/_protected/database/$id/table/-components/table/table-header-cell'
+import { TableInfiniteLoader } from '~/routes/_protected/database/$id/table/-components/table/table-infinite-loader'
+import { TableBodySkeleton } from '~/routes/_protected/database/$id/table/-components/table/table-skeleton'
 import { getColumnSize } from '~/routes/_protected/database/$id/table/-lib'
 import { useTableColumns } from '~/routes/_protected/database/$id/table/-queries/use-columns-query'
 import { TableCellContent } from './table-cell-content'
