+++ apps/desktop/src/routes/_protected/database/$id/table/-components/header/header-actions.tsx
@@ -10,8 +10,8 @@ import { ExportData } from '~/components/export-data'
 import { connectionConstraintsQuery, connectionRowsQuery, connectionTableColumnsQuery, connectionTableTotalQuery } from '~/entities/connection/queries'
 import { rowsQuery } from '~/entities/connection/sql/rows'
 import { queryClient } from '~/main'
-import { Route } from '..'
-import { usePageStoreContext } from '../-store'
+import { Route } from '../../'
+import { usePageStoreContext } from '../../-store'
 import { HeaderActionsColumns } from './header-actions-columns'
 import { HeaderActionsDelete } from './header-actions-delete'
 import { HeaderActionsFilters } from './header-actions-filters'
