+++ apps/desktop/src/routes/_protected/database/$id/table/-components/header/header.tsx
@@ -2,9 +2,9 @@ import { Separator } from '@conar/ui/components/separator'
 import NumberFlow from '@number-flow/react'
 import { useStore } from '@tanstack/react-store'
 import { useConnectionTableTotal } from '~/entities/connection/queries'
-import { Route } from '..'
-import { useTableColumns } from '../-queries/use-columns-query'
-import { usePageStoreContext } from '../-store'
+import { Route } from '../..'
+import { useTableColumns } from '../../-queries/use-columns-query'
+import { usePageStoreContext } from '../../-store'
 import { HeaderActions } from './header-actions'
 import { HeaderSearch } from './header-search'
