+++ apps/desktop/src/routes/_protected/database/$id/table/-components/table/rename-column-dialog.tsx
