+++ apps/desktop/src/routes/_protected/database/$id/table/-components/table/table-empty.tsx
