+++ apps/desktop/src/routes/_protected/database/$id/table/-components/table/table.tsx
@@ -13,10 +13,11 @@ import { connectionRowsQuery } from '~/entities/connection/queries'
 import { useConnectionEnums } from '~/entities/connection/queries/enums'
 import { selectQuery, setQuery } from '~/entities/connection/sql'
 import { queryClient } from '~/main'
-import { Route } from '..'
-import { getColumnSize, selectSymbol } from '../-lib'
-import { useTableColumns } from '../-queries/use-columns-query'
-import { usePageStoreContext } from '../-store'
+import { Route } from '../..'
+import { getColumnSize, selectSymbol } from '../../-lib'
+import { useTableColumns } from '../../-queries/use-columns-query'
+import { usePageStoreContext } from '../../-store'
+import { useColumnsOrder } from '../use-columns-order'
 import { RenameColumnDialog } from './rename-column-dialog'
 import { TableEmpty } from './table-empty'
 import { TableHeader } from './table-header'
@@ -24,7 +25,6 @@ import { TableHeaderCell } from './table-header-cell'
 import { TableInfiniteLoader } from './table-infinite-loader'
 import { SelectionCell, SelectionHeaderCell } from './table-selection'
 import { TableBodySkeleton } from './table-skeleton'
-import { useHeaderActionsOrder } from './use-header-actions-order'
 
 function prepareValue(value: unknown, column?: Column): unknown {
   if (!column)
@@ -66,7 +66,7 @@ function TableComponent({ table, schema }: { table: string, schema: string }) {
   const [filters, orderBy] = useStore(store, state => [state.filters, state.orderBy])
   const { data: rows, error, isPending: isRowsPending } = useInfiniteQuery(connectionRowsQuery({ connection, table, schema, query: { filters, orderBy } }))
   const primaryColumns = useMemo(() => columns?.filter(c => c.primaryKey).map(c => c.id) ?? [], [columns])
-  const { onOrder } = useHeaderActionsOrder()
+  const { toggleOrder } = useColumnsOrder()
   const renameColumnRef = useRef<{ rename: (schema: string, table: string, column: string) => void }>(null)
 
   useEffect(() => {
@@ -217,7 +217,7 @@ function TableComponent({ table, schema }: { table: string, schema: string }) {
           return (
             <TableHeaderCell
               column={column}
-              onSort={() => onOrder(column.id)}
+              onSort={() => toggleOrder(column.id)}
               onRename={onRename}
               {...props}
             />
@@ -250,7 +250,7 @@ function TableComponent({ table, schema }: { table: string, schema: string }) {
     }
 
     return sortedColumns
-  }, [connection, table, schema, columns, hiddenColumns, primaryColumns, saveValue, onOrder, enums])
+  }, [connection, table, schema, columns, hiddenColumns, primaryColumns, saveValue, toggleOrder, enums])
 
   return (
     <TableProvider
