+++ apps/desktop/src/routes/_protected/database/$id/table/index.tsx
@@ -13,9 +13,9 @@ import { FiltersProvider } from '~/components/table'
 import { addTab, connectionStore } from '~/entities/connection/store'
 import { prefetchConnectionCore, prefetchConnectionTableCore } from '~/entities/connection/utils'
 import { Filters } from './-components/filters'
-import { Header } from './-components/header'
+import { Header } from './-components/header/header'
 import { Sidebar } from './-components/sidebar'
-import { Table } from './-components/table'
+import { Table } from './-components/table/table'
 import { TablesTabs } from './-components/tabs'
 import { useTableColumns } from './-queries/use-columns-query'
 import { createPageStore, PageStoreContext } from './-store'
